+++ CMake/PortAudioConfig.cmake
@@ -24,10 +24,15 @@
 # ==============================================================================
 if(PORTAUDIO_FOUND)
   # workaround for github CI with catalina
-  if(EXISTS "/usr/local/Cellar/portaudio/19.6.0/include")
+  if(EXISTS "/usr/local/Cellar/portaudio/19.7.0/include")
+    set(PORTAUDIO_INCLUDE_DIRS /usr/local/Cellar/portaudio/19.7.0/include)
+  elseif(EXISTS "/usr/local/Cellar/portaudio/19.6.0/include")
     set(PORTAUDIO_INCLUDE_DIRS /usr/local/Cellar/portaudio/19.6.0/include)
   endif()
-  if(EXISTS "/usr/local/Cellar/portaudio/19.6.0/include")
+  if(EXISTS "/usr/local/Cellar/portaudio/19.7.0/include")
+    set(PORTAUDIO_LIBRARIES
+        /usr/local/Cellar/portaudio/19.7.0/lib/libportaudio.dylib)
+  elseif(EXISTS "/usr/local/Cellar/portaudio/19.6.0/include")
     set(PORTAUDIO_LIBRARIES
         /usr/local/Cellar/portaudio/19.6.0/lib/libportaudio.dylib)
   endif()