+++ modules/types/builtin/cpp/isfloatBuiltin.cpp
@@ -39,6 +39,13 @@ Nelson::TypeGateway::isfloatBuiltin(Evaluator* eval, int nLhs, const ArrayOfVect
     if (eval->mustOverloadBasicTypes()) {
         retval = OverloadFunction(eval, nLhs, argIn, "isfloat", bSuccess);
     }
+    if (argIn[0].isClassStruct() || argIn[0].isHandle()) {
+        bool bSuccess = false;
+        retval = OverloadFunction(eval, nLhs, argIn, "isfloat", bSuccess);
+        if (bSuccess) {
+            return retval;
+        }
+    }
     if (!bSuccess) {
         bool bRes = (argIn[0].getDataClass() == NLS_DOUBLE
             || argIn[0].getDataClass() == NLS_DCOMPLEX || argIn[0].getDataClass() == NLS_SINGLE