+++ modules/types/builtin/cpp/isnumericBuiltin.cpp
@@ -39,6 +39,13 @@ Nelson::TypeGateway::isnumericBuiltin(Evaluator* eval, int nLhs, const ArrayOfVe
     if (eval->mustOverloadBasicTypes()) {
         retval = OverloadFunction(eval, nLhs, argIn, "isnumeric", bSuccess);
     }
+    if (argIn[0].isClassStruct() || argIn[0].isHandle()) {
+        bool bSuccess = false;
+        retval = OverloadFunction(eval, nLhs, argIn, "isnumeric", bSuccess);
+        if (bSuccess) {
+            return retval;
+        }
+    }
     if (!bSuccess) {
         retval << ArrayOf::logicalConstructor(argIn[0].isNumeric());
     }