+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## Features:
 
+- Validators functions (part 1) available from Nelson and C++ API:
+
+  - `mustBeLogicalScalar`, `mustBeLogical`, `mustBeFloat`, 
+  - `mustBeFinite`, `mustBeScalarOrEmpty`, `mustBeVector`, 
+  - `mustBeValidVariableName`,
+  - `mustBeTextScalar`, `mustBeFolder`,
+  - `mustBeNumeric`.
+
+- extends `assert_checkerror` to check also error identifier.
+
+- `isvector`, `isscalar` support overload.
+
+- `isvarname` builtin: check if input is valid variable name.
+
+- `isdir` manages string array.
+
+- `time` returns current time as the number of seconds or nanoseconds since the epoch.
+
 - Functions using SIMD extensions:
 
   - `ceil`, `round`, `fix`, `floor`, `abs`, `conj`,
@@ -9,6 +27,10 @@
   - `cos`, `sin`, `tan`
   - `atan2`, `acos`, `asin`
   - addition, substraction, multiplication, division vectors.
+
+- `system` allows to run shell command execution in parallel.
+
+- `test_run` executes on parallel process.
 
 - [#352](http://github.com/Nelson-numerical-software/nelson/issues/352): number of input arguments checked in macro.